requirements.txt
Flask
weaviate-client
Flask
pandas
numpy
scikit-learn
cohere